fix: screensaver bugs, move controller routes to own file
